requirements.txt
flask
line-bot-sdk
groq
pandas
openpyxl
python-dotenv
gunicorn
numpy
